refactor: expose Validate function
